Refactor architecture per issue #86 — controller, platform adapters, pure parseInput
- Extract useTaxAppController hook from App.jsx; App.jsx is now a ~130-line composition shell
- Add src/application/parseInput.js: pure use-case entrypoint accepting plain text/DOM, no File APIs
- Add src/platform/web/fileReader.js: browser-side File reading adapter, delegates to parseInput
- Add src/platform/web/htmlParser.js: wraps DOMParser so domain stays browser-free
- Add src/platform/web/themeStorage.js: wraps localStorage + document.documentElement
- Add src/hooks/useThemeMode.js: day/night toggle using themeStorage adapter
- Remove src/application/readInput.js (replaced by platform/web/fileReader.js)
- Update parseTradesFromHtml to accept a pre-parsed Document (DOMParser removed from domain)
- Update parseTradeConfirmationHtml to accept a Document (DOMParser removed from parsing layer)
- Update CLAUDE.md to document final layer structure, dependency direction, and i18n change

Closes #86
