+++ src/domain/parser/__tests__/parseTradesHtml.test.js
@@ -2,6 +2,10 @@
 import { describe, it, expect } from 'vitest'
 import { parseTradesFromHtml } from '../parseTradesHtml.js'
 
+function parseHtml(html) {
+  return new DOMParser().parseFromString(html, 'text/html')
+}
+
 // All tbodies must live inside a <table> or the HTML parser hoists them out.
 function buildHtml(tableContent) {
   return `<!DOCTYPE html><html><body>
@@ -67,17 +71,17 @@ function trade(fields) {
 
 describe('parseTradesFromHtml', () => {
   it('returns empty array for empty document', () => {
-    expect(parseTradesFromHtml(buildHtml(''))).toEqual([])
+    expect(parseTradesFromHtml(parseHtml(buildHtml('')))).toEqual([])
   })
 
   it('returns empty array when only summary tbodies exist (no row-detail)', () => {
     const html = buildHtml(stocksSection('USD', summaryTbody({})))
-    expect(parseTradesFromHtml(html)).toEqual([])
+    expect(parseTradesFromHtml(parseHtml(html))).toEqual([])
   })
 
   it('parses a single BUY trade from a detail tbody', () => {
     const html = buildHtml(stocksSection('USD', trade({})))
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result).toHaveLength(1)
     expect(result[0]).toMatchObject({
       asset:      'Stocks',
@@ -102,7 +106,7 @@ describe('parseTradesFromHtml', () => {
       symbol: 'AAPL', side: 'SELL', qty: '-5',
       price: '180.0000', proceeds: '900.00', code: 'C',
     })))
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result).toHaveLength(1)
     expect(result[0].side).toBe('SELL')
     expect(result[0].quantity).toBe('-5')
@@ -121,7 +125,7 @@ describe('parseTradesFromHtml', () => {
 </tr>
 </tbody>`
     const html = buildHtml(stocksSection('USD', leakyDetail))
-    expect(parseTradesFromHtml(html)).toEqual([])
+    expect(parseTradesFromHtml(parseHtml(html))).toEqual([])
   })
 
   it('skips detail rows with fewer than 13 cells', () => {
@@ -130,7 +134,7 @@ describe('parseTradesFromHtml', () => {
 <tr><td>U1</td><td>AAPL</td><td>2025-01-01</td></tr>
 </tbody>`
     const html = buildHtml(stocksSection('USD', shortRow))
-    expect(parseTradesFromHtml(html)).toEqual([])
+    expect(parseTradesFromHtml(parseHtml(html))).toEqual([])
   })
 
   it('skips entire Forex section', () => {
@@ -138,7 +142,7 @@ describe('parseTradesFromHtml', () => {
 <tbody><tr><td class="header-asset">Forex</td></tr></tbody>
 <tbody><tr><td class="header-currency">USD</td></tr></tbody>
 ${trade({ symbol: 'EUR.USD', side: 'BUY' })}`)
-    expect(parseTradesFromHtml(html)).toEqual([])
+    expect(parseTradesFromHtml(parseHtml(html))).toEqual([])
   })
 
   it('skips FX section (alternate asset header text)', () => {
@@ -146,7 +150,7 @@ ${trade({ symbol: 'EUR.USD', side: 'BUY' })}`)
 <tbody><tr><td class="header-asset">FX</td></tr></tbody>
 <tbody><tr><td class="header-currency">USD</td></tr></tbody>
 ${trade({ symbol: 'EUR.USD' })}`)
-    expect(parseTradesFromHtml(html)).toEqual([])
+    expect(parseTradesFromHtml(parseHtml(html))).toEqual([])
   })
 
   it('parses multiple trades under the same currency', () => {
@@ -155,7 +159,7 @@ ${trade({ symbol: 'EUR.USD' })}`)
       trade({ symbol: 'AAPL', side: 'SELL', qty: '-5', code: 'C' }) +
       trade({ symbol: 'GOOG', side: 'BUY', qty: '2' })
     ))
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result).toHaveLength(3)
     expect(result.map(r => r.symbol)).toEqual(['AAPL', 'AAPL', 'GOOG'])
   })
@@ -165,7 +169,7 @@ ${trade({ symbol: 'EUR.USD' })}`)
       trade({ symbol: 'VWCE' }) +
       trade({ symbol: 'IWDA' })
     ))
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result).toHaveLength(2)
     expect(result.every(r => r.currency === 'EUR')).toBe(true)
   })
@@ -177,7 +181,7 @@ ${trade({ symbol: 'EUR.USD' })}`)
 ${trade({ symbol: 'AAPL' })}
 <tbody><tr><td class="header-currency">EUR</td></tr></tbody>
 ${trade({ symbol: 'VWCE' })}`)
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result).toHaveLength(2)
     expect(result[0].currency).toBe('USD')
     expect(result[1].currency).toBe('EUR')
@@ -193,7 +197,7 @@ ${trade({ symbol: 'EUR.USD' })}
 <tbody><tr><td class="header-asset">Stocks</td></tr></tbody>
 <tbody><tr><td class="header-currency">EUR</td></tr></tbody>
 ${trade({ symbol: 'VWCE' })}`)
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result).toHaveLength(2)
     expect(result.map(r => r.symbol)).toEqual(['AAPL', 'VWCE'])
   })
@@ -204,12 +208,12 @@ ${trade({ symbol: 'VWCE' })}`)
 <td>10</td><td>174.5</td><td>-1745</td><td>-2.5</td><td>0</td><td>&nbsp;</td><td>&nbsp;</td>
 </tr></tbody>`
     const html = buildHtml(stocksSection('USD', trade({}) + subtotal))
-    expect(parseTradesFromHtml(html)).toHaveLength(1)
+    expect(parseTradesFromHtml(parseHtml(html))).toHaveLength(1)
   })
 
   it('returns asset field from the current Stocks header', () => {
     const html = buildHtml(stocksSection('USD', trade({})))
-    const result = parseTradesFromHtml(html)
+    const result = parseTradesFromHtml(parseHtml(html))
     expect(result[0].asset).toBe('Stocks')
   })
 })
